Update Webnovel_Downloader_PyQt6_UI.py
Fixed bug where app crashes in automatic mode when too many profiles are opted in. Introduced a 3 second delay.

Added the option to send created epubs to a discord webhook instead of a local directory. In the future, I'm planning on improving the input for it in the UI and saving the webhook URL locally so it remembers it when re-launching the application, but that's in the future.

Got rid of "fetch all chapters" toggle, and instead added a "load next 50 chapters" and a "load all chapters" button in the download preview.

One-click download and automatic mode only fetch new chapters, but will now only fetch a maximum of 50 chapters. Use download preview for more chapters.